+++ Vault/src/utopia/vault/nosql/view/DeprecatableView.scala
@@ -1,6 +1,5 @@
 package utopia.vault.nosql.view
 
-import utopia.vault.database.Connection
 import utopia.vault.model.template.Deprecates
 
 /**
@@ -17,13 +16,6 @@ trait DeprecatableView[+Repr] extends FilterableView[Repr]
 	 * @return Model used for interacting with the DB and for building conditions
 	 */
 	protected def model: Deprecates
-	
-	/**
-	 * Deprecates the accessible items
-	 * @param connection Implicit DB connection
-	 * @return Whether any row was targeted
-	 */
-	def deprecate()(implicit connection: Connection): Boolean
 	
 	
 	// COMPUTED ----------------------